@@ -8,22 +8,22 @@
         "@testing-library/jest-dom": "^5.16.4",
         "@testing-library/react": "^13.3.0",
         "@testing-library/user-event": "^14.2.1",
-        "@types/jest": "^27.5.2",
+        "@types/jest": "^28.1.3",
         "@types/node": "^16.11.41",
-        "@types/react": "^17.0.47",
-        "@types/react-dom": "^17.0.17",
+        "@types/react": "^18.0.14",
+        "@types/react-dom": "^18.0.5",
         "@types/styled-components": "^5.1.25",
         "@types/uuid": "^8.3.4",
         "aubiojs": "^0.0.1",
         "browserify-fs": "^1.0.0",
         "buffer": "^6.0.3",
-        "gh-pages": "^3.2.3",
+        "gh-pages": "^4.0.0",
         "howler": "^2.2.3",
         "lodash": "^4.17.21",
         "path-browserify": "^1.0.1",
         "peerjs": "^2.0.0-beta.3",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-hotkeys-hook": "^3.4.6",
         "react-query": "^3.39.1",
         "react-scripts": "5.0.1",
